+++ dse_do_utils/datamanager.py
@@ -347,21 +347,54 @@ class DataManager(object):
     #             lambda row: pd.Series(func(row, **kwargs), index=column_names), axis=1)), axis=1, sort=False)
 
     @staticmethod
-    def extract_solution(df: object, extract_dvar_names: List[str] = None, drop_column_names: List[str] = None,
-                         drop: bool = True) -> object:
+    def extract_solution(df: pd.DataFrame, extract_dvar_names: List[str] = None, drop_column_names: List[str] = None,
+                         drop: bool = True, epsilon: float = None) -> pd.DataFrame:
         """Generalized routine to extract a solution value.
-        Can remove the dvar column from the df to be able to have a clean df for export into scenario."""
+        Can remove the dvar column from the df to be able to have a clean df for export into scenario.
+
+        In some cases, CPLEX extracted values for continuous dvars can have very small values instead of zeros.
+        If epsilon has a value, this method will drop these small values to zero.
+        And it will assume the values need to be positive, so it clips negative values at zero.
+
+        Args:
+            df: DataFrame
+            extract_dvar_names: list of column names with CPLEX dvars
+            drop_column_names: columns to be dropped (can be different and in addition to drop)
+            drop: if True drops all columns in extract_dvar_names
+            epsilon (float): if not None, drop values below threshold to zero and clip negative values at zero
+
+        """
         if extract_dvar_names is not None:
             for xDVarName in extract_dvar_names:
                 if xDVarName in df.columns:
-                    df[f'{xDVarName}Sol'] = [dvar.solution_value for dvar in df[xDVarName]]
+                    solution_column_name = f'{xDVarName}Sol'
+                    df[solution_column_name] = [dvar.solution_value for dvar in df[xDVarName]]
                     if drop:
                         df = df.drop([xDVarName], axis=1)
+                    if epsilon is not None:
+                        df.loc[df[solution_column_name] < epsilon, solution_column_name] = 0
+                        df[solution_column_name] = df[solution_column_name].clip(lower=0)
         if drop and drop_column_names is not None:
             for column in drop_column_names:
                 if column in df.columns:
                     df = df.drop([column], axis=1)
         return df
+
+    # def drop_small_epsilon_values(self, df: pd.DataFrame, columns: List[str], epsilon: float = 0.0001) -> pd.DataFrame:
+    #     """Drops small values of extracted CPLEX continuous dvar solutions to zero.
+    #     In some cases, CPLEX extracted values can have very small values instead of zeros.
+    #     This method drops these small values to zero.
+    #     It also assumes the values need to be positive, so it clips all negative values at zero.
+    #     Args:
+    #         df: DataFrame
+    #         columns: List of column names
+    #         epsilon: threshold value
+    #     """
+    #     for col in columns:
+    #         if col in df.columns:
+    #             df.loc[df[col] < epsilon, col] = 0
+    #             df[col] = df[col].clip(lower=0)
+    #     return df
 
     def get_raw_table_by_name(self, table_name: str) -> Optional[pd.DataFrame]:
         """Get the 'raw' (non-indexed) table from inputs or outputs."""