+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Core01EnvironmentManager.find_project_root_directory
 - MapManager.add_bar_chart_in_map
+- DataManager.extract_solution adds option to drop small values to zero
 ### Changed
 - Core01DataManager.prepare_df now merges local dtypes with self.dtypes, where overlapping local dtypes take priority